sap_core_next/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sap-core-next"
version = "0.1.0"
description = "Agnostic SAP automation engine with plugin architecture"
requires-python = ">=3.10"
dependencies = [
  "openpyxl>=3.1",
  "pyyaml>=6.0",
  "requests>=2.31",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
